requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0

# Text-to-speech (optional - install one)
pyttsx3>=2.90
# gtts>=2.4.0  # Alternative TTS (requires internet)

# Audio playback for gTTS
# scipy>=1.11.0  # For saving audio files
